\name{summary.som}
\alias{summary.som}
\alias{print.som}
\title{ summarize a som object }
\description{print out the configuration parameters of a som object
}
\usage{
\method{summary}{som}(object, \dots)
\method{print}{som}(x, \dots)
}

\arguments{
  \item{object, x}{ a `som' object }
  \item{\dots}{nothing yet}
}

\author{
  Jun Yan <jyan@stat.uiowa.edu>
}

\keyword{print}

